+++ src/vue/modules/create-sale-form/mixins/manage-sale-request.mixin.js
@@ -39,11 +39,13 @@ export default {
         softCap: this.informationStepForm.softCap,
         hardCap: this.informationStepForm.hardCap,
         requiredBaseAssetForHardCap: this.informationStepForm.assetsToSell,
-        quoteAssets: this.informationStepForm.quoteAssets
-          .map((item) => ({
-            asset: item,
-            price: DEFAULT_QUOTE_ASSET_PRICE,
-          })),
+        quoteAssets: [
+          ...this.informationStepForm.quoteAssets,
+          this.informationStepForm.capAsset.code,
+        ].map((item) => ({
+          asset: item,
+          price: DEFAULT_QUOTE_ASSET_PRICE,
+        })),
         creatorDetails: {
           name: this.informationStepForm.name,
           short_description: this.shortBlurbStepForm.shortDescription,