+++ src/vue/modules/movements-top-bar/index.vue
@@ -10,11 +10,18 @@
             {{ 'op-pages.filters-prefix' | globalize }}
           </span>
           <select-field
-            v-model="asset"
-            :values="assets"
-            key-as-value-text="nameAndCode"
+            :value="asset.code"
+            @input="setAssetByCode"
             class="app__select app__select--no-border"
-          />
+          >
+            <option
+              v-for="asset in assets"
+              :key="asset.code"
+              :value="asset.code"
+            >
+              {{ asset.nameAndCode }}
+            </option>
+          </select-field>
         </div>
       </template>
       <div
@@ -186,6 +193,9 @@ export default {
     ...mapActions({
       loadAccountBalancesDetails: vuexTypes.LOAD_ACCOUNT_BALANCES_DETAILS,
     }),
+    setAssetByCode (code) {
+      this.asset = this.assets.find(item => item.code === code)
+    },
     setDefaultAsset () {
       this.asset = this.assets
         .find(item => item.code === this.$route.query.asset) ||